rename persistence file in coord to cache
